requirements.txt
fastapi
uvicorn
timm
pillow
python-multipart
torch==2.5.1+cpu
torchvision==0.20.1+cpu